Separate `skydb/pq/oauth.go` from `skydb/pq/user.go`
